Update library structure
