+++ lib/src/trie_router/src/router_result.dart
@@ -1,4 +1,4 @@
-import '../../../routemaster.dart';
+import '../../core.dart';
 
 /// A match for a path returned from [TrieRouter].
 class RouterResult {
